ai-services/requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6
opencv-python==4.9.0.80
numpy==1.26.3
Pillow==10.2.0
rembg==2.0.56
tensorflow==2.15.0
torch==2.1.2
torchvision==0.16.2
ultralytics==8.0.230
requests==2.31.0
python-dotenv==1.0.0
aiofiles==23.2.1
